refactor(citation-validator): convert path-resolution to strategy array and split types (closes #28)
- Extract pathResolutionStrategies/ matching eligibilityStrategies pattern
- Refactor validateCrossDocumentLink to iterate strategy array (no waterfall)
- Split contentExtractorTypes.ts into cli-types/extraction-types/strategy-types
- CitationValidator reduced from 654 to 397 lines
